Move JobsOperation->operation and JobOperation->legacy.
And update JobsOperation signature.
