+++ flow/__init__.py
@@ -13,8 +13,8 @@ from . import environment
 from . import scheduling
 from . import errors
 from .project import FlowProject
-from .project import JobOperation
-from .project import JobsOperation
+from .legacy import JobOperation
+from .operations import JobsOperation
 from .project import label
 from .project import classlabel
 from .project import staticlabel
