+++ flow/operations.py
@@ -12,15 +12,19 @@ import argparse
 import logging
 import inspect
 from multiprocessing import Pool
+from hashlib import sha1
 from functools import wraps
 from itertools import groupby
 from itertools import zip_longest
 
 from signac import get_project
 from signac.common import six
+from signac.contrib.hashing import calc_id
 
+from .scheduling.base import JobStatus
 from .util.tqdm import tqdm
 from .util.execution import fork
+from .util.misc import TrackGetItemDict
 if six.PY2:
     from collections import Iterable
 else:
@@ -328,6 +332,146 @@ def run(parser=None):
             result = pool.imap_unordered(operation, jobs)
             for _ in tqdm(jobs) if args.progress else jobs:
                 result.next(args.timeout)
+
+
+class JobsOperation(object):
+    """This class represents the information needed to execute one operation for jobs.
+
+    An operation function in this context is a shell command, which should be a function
+    of one or multiple signac jobs.
+
+    .. note::
+
+        This class is used by the :class:`~.FlowProject` class for the execution and
+        submission process and should not be instantiated by users themselves.
+
+    :param name:
+        The name of this JobsOperation instance. The name is arbitrary,
+        but helps to concisely identify the operation in various contexts.
+    :type name:
+        str
+    :param cmd:
+        The command that executes this operation.
+    :type cmd:
+        str
+    :param jobs:
+        The sequence of jobs associated with this operation.
+    :type job:
+        :py:class:`signac.Job`.
+    :param directives:
+        A dictionary of additional parameters that provide instructions on how
+        to execute this operation, e.g., specifically required resources.
+    :type directives:
+        :class:`dict`
+    """
+    MAX_LEN_ID = 100
+
+    def __init__(self, name, cmd, jobs, directives=None):
+        if not len(jobs):
+            raise ValueError("The jobs argument cannot be empty!")
+        self.name = name
+        self.jobs = jobs
+        self.cmd = cmd
+        if directives is None:
+            directives = dict()  # default argument
+        else:
+            directives = dict(directives)  # explicit copy
+
+        # Keys which were explicitly set by the user, but are not evaluated by the
+        # template engine are cause for concern and might hint at a bug in the template
+        # script or ill-defined directives. We are therefore keeping track of all
+        # keys set by the user and check whether they have been evaluated by the template
+        # script engine later.
+        keys_set_by_user = set(directives.keys())
+
+        # Evaluate strings and callables for job:
+        def evaluate(value):
+            if value and callable(value):
+                return value(* self.jobs)
+            elif isinstance(value, six.string_types):
+                if len(jobs) == 1:
+                    return value.format(job=self.jobs[0])
+                else:
+                    return value.format(jobs=self.jobs)
+            else:
+                return value
+
+        directives.setdefault('np',
+                              evaluate(directives.get('nranks', 1))
+                              * evaluate(directives.get('omp_num_threads', 1)))
+        directives.setdefault('ngpu', 0)
+        directives.setdefault('nranks', 0)
+        directives.setdefault('omp_num_threads', 0)
+
+        # We use a special dictionary that allows us to track all keys that have been
+        # evaluated by the template engine and compare them to those explicitly set
+        # by the user. See also comment above.
+        self.directives = TrackGetItemDict(
+            {key: evaluate(value) for key, value in directives.items()})
+        self.directives._keys_set_by_user = keys_set_by_user
+
+    def __str__(self):
+        return "{}({})".format(self.name, ', '.join([job.get_id() for job in self.jobs]))
+
+    def __repr__(self):
+        return "{type}(name='{name}', cmd={cmd}, jobs='{jobs}', directives={directives})".format(
+            type=type(self).__name__,
+            name=self.name,
+            jobs=', '.join([job.get_id() for job in self.jobs]),
+            cmd=repr(self.cmd),
+            directives=self.directives)
+
+    @property
+    def job(self):
+        assert len(self.jobs) >= 1
+        return self.jobs[0]
+
+    def get_id(self, index=0):
+        "Return a name, which identifies this job-operation."
+        project = self.jobs[0]._project
+
+        # The full name is designed to be truly unique for each job-operation.
+        full_name = '{}%{}%{}%{}'.format(
+            project.root_directory(),
+            ','.join([job.get_id() for job in self.jobs]),
+            self.name, index)
+
+        # The jobs_op_id is a hash computed from the unique full name.
+        jobs_op_id = calc_id(full_name)
+
+        # The actual job id is then constructed from a readable part and the jobs_op_id,
+        # ensuring that the job-op is still somewhat identifiable, but guarantueed to
+        # be unique. The readable name is based on the project id, job id, operation name,
+        # and the index number. All names and the id itself are restricted in length
+        # to guarantuee that the id does not get too long.
+        max_len = self.MAX_LEN_ID - len(jobs_op_id)
+        if max_len < len(jobs_op_id):
+            raise ValueError("Value for MAX_LEN_ID is too small ({}).".format(self.MAX_LEN_ID))
+
+        readable_name = '{}/{}/{}/{:04d}/'.format(
+            str(project)[:12], ','.join([str(job)[:8] for job in self.jobs]),
+            self.name[:12], index)[:max_len]
+
+        # By appending the unique jobs_op_id, we ensure that each id is truly unique.
+        return readable_name + jobs_op_id
+
+    def __hash__(self):
+        return int(sha1(self.get_id().encode('utf-8')).hexdigest(), 16)
+
+    def __eq__(self, other):
+        return self.get_id() == other.get_id()
+
+    def set_status(self, value):
+        "Store the operation's status."
+        self.job._project.document.setdefault('_status', dict())
+        self.job._project.document._status[self.get_id()] = int(value)
+
+    def get_status(self):
+        "Retrieve the operation's last known status."
+        try:
+            return JobStatus(self.job._project.document['_status'][self.get_id()])
+        except KeyError:
+            return JobStatus.unknown
 
 
 __all__ = ['cmd', 'directives', 'run']
